requirements.txt
langchain
python-dotenv
streamlit
langchain_openai 
langchain_core
langchain_community
langserve
fastapi
uvicorn